requirements.txt
flask
certifi==2022.9.14
charset-normalizer==2.1.1
colorama==0.4.5
customtkinter==4.6.1
darkdetect==0.7.1
diffusers==0.20.0
filelock==3.8.0
huggingface-hub==0.20.3
idna==3.4
importlib-metadata==4.12.0
numpy==1.23.3
packaging==21.3
Pillow==9.2.0
pyparsing==3.0.9
PyYAML==6.0
regex==2022.9.13
requests==2.28.1
tk==0.1.0
tokenizers==0.15.2
tqdm==4.64.1
transformers==4.39.3
typing_extensions==4.8.0
urllib3==1.26.12
zipp==3.8.1